backend/pyproject.toml
[project]
name = "voice-assistant-backend"
version = "0.1.0"
description = "FastAPI backend for managing AI voice assistants, templates, and voice profiles."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "sqlmodel>=0.0.20,<0.0.21",
    "asyncpg>=0.29.0,<0.30.0",
    "alembic>=1.13.0,<2.0.0",
    "pydantic-settings>=2.3.0,<3.0.0",
    "httpx>=0.28.1,<1.0.0",
    "structlog>=24.1.0,<25.0.0",
    "twilio>=9.7.2,<10.0.0",
    "websockets>=15.0.1,<16.0.0",
]
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "ruff>=0.6.0,<0.7.0",
    "mypy>=1.11.0,<2.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B"]
extend-ignore = ["E203"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
strict = true
plugins = ["pydantic.mypy"]

[tool.pdm]
distribution = false
